Another big rewrite (#19)

+++ banchus-common/src/main/kotlin/pe/nanamochi/banchus/domain/enums/SlotTeam.kt
@@ -1,11 +1,11 @@
 package pe.nanamochi.banchus.domain.enums
 
-enum class SlotTeam(val value: Int) {
-    NEUTRAL(0),
-    BLUE(1),
-    RED(2);
+enum class SlotTeam(val value: UByte) {
+    NEUTRAL(0u),
+    BLUE(1u),
+    RED(2u);
 
     companion object {
-        fun fromValue(value: Int): SlotTeam = entries.find { it.value == value } ?: NEUTRAL
+        fun fromValue(value: UByte): SlotTeam = entries.find { it.value == value } ?: NEUTRAL
     }
 }
